refactor(core): split render-svg-scene chrome and legend SVG
Close #950. Extract panel chrome (axes/grid/strips/titles) and legend
SVG (discrete/steps/ramp) from the scene assembly so each file has one
primary responsibility. Public sceneLabel / sceneToSVGString exports
and SVG markup stay unchanged.
